refactor: adapter/in 패키지 구조 변경 - onboarding에서 mission 분리
