+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/mission/dto/MissionListResponse.java
@@ -1,4 +1,4 @@
-package com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto;
+package com.booquest.booquest_api.adapter.in.mission.dto;
 
 import lombok.Builder;
 import lombok.Getter;
