+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/mission/dto/MissionResponseDto.java
@@ -1,4 +1,4 @@
-package com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto;
+package com.booquest.booquest_api.adapter.in.mission.dto;
 
 import com.booquest.booquest_api.domain.mission.model.Mission;
