+++ booquest-api/src/main/java/com/booquest/booquest_api/adapter/in/onboarding/web/dto/SideJobDetailResponseDto.java
@@ -1,8 +1,8 @@
 package com.booquest.booquest_api.adapter.in.onboarding.web.dto;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionDetailResponseDto;
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionResponseDto;
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.missionstep.dto.MissionStepResponseDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionDetailResponseDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionResponseDto;
+import com.booquest.booquest_api.adapter.in.mission.missionstep.dto.MissionStepResponseDto;
 import com.booquest.booquest_api.domain.sidejob.model.SideJob;
 
 import java.util.List;
