+++ booquest-api/src/main/java/com/booquest/booquest_api/application/port/in/mission/GetMissionListUseCase.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.application.port.in.mission;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionResponse;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionResponse;
 import com.booquest.booquest_api.domain.sidejob.enums.MissionStatus;
 
 import java.util.List;
