+++ booquest-api/src/main/java/com/booquest/booquest_api/application/port/in/sidejob/mission/GenerateMissionUseCase.java
@@ -1,7 +1,6 @@
 package com.booquest.booquest_api.application.port.in.sidejob.mission;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.dto.SideJobResponseDto;
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionGenerateRequestDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionGenerateRequestDto;
 import com.booquest.booquest_api.domain.mission.model.Mission;
 
 import java.util.List;
