+++ booquest-api/src/main/java/com/booquest/booquest_api/application/port/in/sidejob/mission/GetMissionProgressUseCase.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.application.port.in.sidejob.mission;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionProgressResponseDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionProgressResponseDto;
 
 public interface GetMissionProgressUseCase {
     MissionProgressResponseDto getMissionProgress(Long userId, Long sideJobId);
