+++ booquest-api/src/main/java/com/booquest/booquest_api/application/port/in/sidejob/mission/missionstep/GenerateMissionStepUseCase.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.application.port.in.sidejob.mission.missionstep;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.missionstep.dto.MissionStepGenerateRequestDto;
+import com.booquest.booquest_api.adapter.in.mission.missionstep.dto.MissionStepGenerateRequestDto;
 import com.booquest.booquest_api.domain.mission.model.MissionStep;
 
 import java.util.List;
