+++ booquest-api/src/main/java/com/booquest/booquest_api/application/service/mission/GetMissionListService.java
@@ -1,7 +1,7 @@
 package com.booquest.booquest_api.application.service.mission;
 
 import com.booquest.booquest_api.application.port.in.mission.GetMissionListUseCase;
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionResponse;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionResponse;
 import com.booquest.booquest_api.application.port.out.sidejob.mission.MissionRepositoryPort;
 import com.booquest.booquest_api.domain.mission.model.Mission;
 import com.booquest.booquest_api.domain.sidejob.enums.MissionStatus;
