+++ booquest-api/src/main/java/com/booquest/booquest_api/application/service/sidejob/mission/GenerateMissionService.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.application.service.sidejob.mission;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionGenerateRequestDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionGenerateRequestDto;
 import com.booquest.booquest_api.application.port.in.sidejob.mission.GenerateMissionResult;
 import com.booquest.booquest_api.application.port.in.sidejob.mission.GenerateMissionUseCase;
 import com.booquest.booquest_api.application.port.out.sidejob.SideJobRepositoryPort;
