+++ booquest-api/src/main/java/com/booquest/booquest_api/application/service/sidejob/mission/GetMissionProgressService.java
@@ -1,6 +1,6 @@
 package com.booquest.booquest_api.application.service.sidejob.mission;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionProgressResponseDto;
+import com.booquest.booquest_api.adapter.in.mission.dto.MissionProgressResponseDto;
 import com.booquest.booquest_api.application.port.in.sidejob.mission.GetMissionProgressUseCase;
 import com.booquest.booquest_api.application.port.out.sidejob.mission.MissionRepositoryPort;
 import com.booquest.booquest_api.application.port.out.sidejob.mission.missionstep.MissionStepRepositoryPort;
