+++ booquest-api/src/main/java/com/booquest/booquest_api/application/service/sidejob/mission/missionstep/GenerateMissionStepService.java
@@ -1,8 +1,6 @@
 package com.booquest.booquest_api.application.service.sidejob.mission.missionstep;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.dto.MissionGenerateRequestDto;
-import com.booquest.booquest_api.adapter.in.onboarding.web.mission.missionstep.dto.MissionStepGenerateRequestDto;
-import com.booquest.booquest_api.application.port.in.sidejob.mission.GenerateMissionResult;
+import com.booquest.booquest_api.adapter.in.mission.missionstep.dto.MissionStepGenerateRequestDto;
 import com.booquest.booquest_api.application.port.in.sidejob.mission.missionstep.GenerateMissionStepUseCase;
 import com.booquest.booquest_api.application.port.out.sidejob.mission.missionstep.GenerateMissionStepPort;
 import com.booquest.booquest_api.application.port.out.sidejob.mission.missionstep.MissionStepRepositoryPort;
